Redirect pip install message from stdout to stderr
It is a convention on linux to output messages that are meant to be
parsed/piped to another program through stdout and diagnostic messages
to stderr. However, due to the fact that many languages only provide
print primitives that work with stdout by default, it is only
understandable that most programs just output to stdout anyway, which is
a shame.